feat(ui): add spotlight-trigger
