+++ apps/console/src/app/app.tsx
@@ -19,6 +19,7 @@ import { useAuthInterceptor } from '@qovery/shared/utils'
 import { environment } from '../environments/environment'
 import PreviewCode from './components/preview-code'
 import ScrollToTop from './components/scroll-to-top'
+import { SpotlightTrigger } from './components/spotlight-trigger'
 import { ROUTER } from './router/main.router'
 
 export function App() {
@@ -126,6 +127,7 @@ export function App() {
         )}
         <Route path="*" element={<Navigate replace to={LOGIN_URL} />} />
       </Routes>
+      <SpotlightTrigger />
     </GTMProvider>
   )
 }
